Workshop 3 Assignment Sdmin

+++ routes/campsiteRouter.js
@@ -16,27 +16,37 @@ campsiteRouter.route('/')
             .catch(err => next(err));
     })
     .post(authenticate.verifyUser, (req, res, next) => {
-        Campsite.create(req.body)
-            .then(campsite => {
-                console.log('Campsite Created ', campsite);
-                res.statusCode = 200;
-                res.setHeader('Content-Type', 'application/json');
-                res.json(campsite);
-            })
-            .catch(err => next(err));
+        if (req.user.admin) {
+            Campsite.create(req.body)
+                .then(campsite => {
+                    console.log('Campsite Created ', campsite);
+                    res.statusCode = 200;
+                    res.setHeader('Content-Type', 'application/json');
+                    res.json(campsite);
+                })
+                .catch(err => next(err));
+        } else {
+            res.statusCode = 403;
+            res.end("You are not authorized to perform this operation!");
+        }
     })
     .put(authenticate.verifyUser, (req, res) => {
         res.statusCode = 403;
         res.end('PUT operation not supported on /campsites');
     })
     .delete(authenticate.verifyUser, (req, res, next) => {
-        Campsite.deleteMany()
-            .then(response => {
-                res.statusCode = 200;
-                res.setHeader('Content-Type', 'application/json');
-                res.json(response);
-            })
-            .catch(err => next(err));
+        if (req.user.admin) {
+            Campsite.deleteMany()
+                .then(response => {
+                    res.statusCode = 200;
+                    res.setHeader('Content-Type', 'application/json');
+                    res.json(response);
+                })
+                .catch(err => next(err));
+        } else {
+            res.statusCode = 403;
+            res.end("You are not authorized to perform this operation!");
+        }
     });
 
 campsiteRouter.route('/:campsiteId')
@@ -56,24 +66,34 @@ campsiteRouter.route('/:campsiteId')
         res.end(`POST operation not supported on /campsites/${req.params.campsiteId}`);
     })
     .put(authenticate.verifyUser, (req, res, next) => {
-        Campsite.findByIdAndUpdate(req.params.campsiteId, {
-            $set: req.body
-        }, { new: true })
-            .then(campsite => {
-                res.statusCode = 200;
-                res.setHeader('Content-Type', 'application/json');
-                res.json(campsite);
-            })
-            .catch(err => next(err));
+        if (req.user.admin) {
+            Campsite.findByIdAndUpdate(req.params.campsiteId, {
+                $set: req.body
+            }, { new: true })
+                .then(campsite => {
+                    res.statusCode = 200;
+                    res.setHeader('Content-Type', 'application/json');
+                    res.json(campsite);
+                })
+                .catch(err => next(err));
+        } else {
+            res.statusCode = 403;
+            res.end("You are not authorized to perform this operation!");
+        }
     })
-    .delete(authenticate.verifyUser, (req, res, next) => {
-        Campsite.findByIdAndDelete(req.params.campsiteId)
-            .then(response => {
-                res.statusCode = 200;
-                res.setHeader('Content-Type', 'application/json');
-                res.json(response);
-            })
-            .catch(err => next(err));
+    .delete(authenticate.verifyUser, (req, res) => {
+        if (req.user.admin) {
+            Campsite.findByIdAndDelete(req.params.campsiteId)
+                .then(response => {
+                    res.statusCode = 200;
+                    res.setHeader('Content-Type', 'application/json');
+                    res.json(response);
+                })
+                .catch(err => next(err));
+        } else {
+            res.statusCode = 403;
+            res.end("You are not authorized to perform this operation!");
+        }
     });
 
 campsiteRouter.route('/:campsiteId/comments')
@@ -93,7 +113,7 @@ campsiteRouter.route('/:campsiteId/comments')
             })
             .catch(err => next(err));
     })
-    .post(authenticate.verifyUser, (req, res, next) => {
+    .post(authenticate.verifyUser, (req, res) => {
         Campsite.findById(req.params.campsiteId)
             .then(campsite => {
                 if (campsite) {
@@ -118,27 +138,32 @@ campsiteRouter.route('/:campsiteId/comments')
         res.statusCode = 403;
         res.end(`PUT operation not supported on /campsites/${req.params.campsiteId}/comments`);
     })
-    .delete(authenticate.verifyUser, (req, res, next) => {
-        Campsite.findById(req.params.campsiteId)
-            .then(campsite => {
-                if (campsite) {
-                    for (let i = (campsite.comments.length - 1); i >= 0; i--) {
-                        campsite.comments.id(campsite.comments[i]._id).remove();
+    .delete(authenticate.verifyUser, (req, res) => {
+        if (req.user.admin) {
+            Campsite.findById(req.params.campsiteId)
+                .then(campsite => {
+                    if (campsite) {
+                        for (let i = (campsite.comments.length - 1); i >= 0; i--) {
+                            campsite.comments.id(campsite.comments[i]._id).remove();
+                        }
+                        campsite.save()
+                            .then(campsite => {
+                                res.statusCode = 200;
+                                res.setHeader('Content-Type', 'application/json');
+                                res.json(campsite);
+                            })
+                            .catch(err => next(err));
+                    } else {
+                        err = new Error(`Campsite ${req.params.campsiteId} not found`);
+                        err.status = 404;
+                        return next(err);
                     }
-                    campsite.save()
-                        .then(campsite => {
-                            res.statusCode = 200;
-                            res.setHeader('Content-Type', 'application/json');
-                            res.json(campsite);
-                        })
-                        .catch(err => next(err));
-                } else {
-                    err = new Error(`Campsite ${req.params.campsiteId} not found`);
-                    err.status = 404;
-                    return next(err);
-                }
-            })
-            .catch(err => next(err));
+                })
+                .catch(err => next(err));
+        } else {
+            res.statusCode = 403;
+            res.end("You are not authorized to perform this operation!");
+        }
     });
 
 campsiteRouter.route('/:campsiteId/comments/:commentId')
@@ -166,23 +191,31 @@ campsiteRouter.route('/:campsiteId/comments/:commentId')
         res.statusCode = 403;
         res.end(`POST operation not supported on /campsites/${req.params.campsiteId}/comments/${req.params.commentId}`);
     })
-    .put(authenticate.verifyUser, (req, res, next) => {
+
+    .put(authenticate.verifyUser, (req, res) => {
         Campsite.findById(req.params.campsiteId)
             .then(campsite => {
                 if (campsite && campsite.comments.id(req.params.commentId)) {
-                    if (req.body.rating) {
-                        campsite.comments.id(req.params.commentId).rating = req.body.rating;
+                    if ((campsite.comments.id(req.params.commentId).author_id).equals(req.user._id)) {
+                        if (req.body.rating) {
+                            campsite.comments.id(req.params.commentId).rating = req.body.rating;
+                        }
+                        if (req.body.text) {
+                            campsite.comments.id(req.params.commentId).text = req.body.text;
+                        }
+                        campsite.save()
+                            .then(campsite => {
+                                res.statusCode = 200;
+                                res.setHeader('Content-Type', 'application/json');
+                                res.json(campsite);
+                            })
+                            .catch(err => next(err));
+                    } else {
+                        const err = new Error("You are not authorized to perform this operation!");
+                        res.statusCode = 403;
+                        return next(err);
+
                     }
-                    if (req.body.text) {
-                        campsite.comments.id(req.params.commentId).text = req.body.text;
-                    }
-                    campsite.save()
-                        .then(campsite => {
-                            res.statusCode = 200;
-                            res.setHeader('Content-Type', 'application/json');
-                            res.json(campsite);
-                        })
-                        .catch(err => next(err));
                 } else if (!campsite) {
                     err = new Error(`Campsite ${req.params.campsiteId} not found`);
                     err.status = 404;
@@ -195,29 +228,31 @@ campsiteRouter.route('/:campsiteId/comments/:commentId')
             })
             .catch(err => next(err));
     })
-    .delete(authenticate.verifyUser, (req, res, next) => {
-        Campsite.findById(req.params.campsiteId)
-            .then(campsite => {
-                if (campsite && campsite.comments.id(req.params.commentId)) {
-                    campsite.comments.id(req.params.commentId).remove();
-                    campsite.save()
-                        .then(campsite => {
-                            res.statusCode = 200;
-                            res.setHeader('Content-Type', 'application/json');
-                            res.json(campsite);
-                        })
-                        .catch(err => next(err));
-                } else if (!campsite) {
-                    err = new Error(`Campsite ${req.params.campsiteId} not found`);
-                    err.status = 404;
-                    return next(err);
-                } else {
-                    err = new Error(`Comment ${req.params.commentId} not found`);
-                    err.status = 404;
-                    return next(err);
-                }
-            })
-            .catch(err => next(err));
+    .delete(authenticate.verifyUser, (req, res) => {
+        if (req.user.admin) {
+            Campsite.findById(req.params.campsiteId)
+                .then(campsite => {
+                    if (campsite && campsite.comments.id(req.params.commentId)) {
+                        campsite.comments.id(req.params.commentId).remove();
+                        campsite.save()
+                            .then(campsite => {
+                                res.statusCode = 200;
+                                res.setHeader('Content-Type', 'application/json');
+                                res.json(campsite);
+                            })
+                            .catch(err => next(err));
+                    } else if (!campsite) {
+                        err = new Error(`Campsite ${req.params.campsiteId} not found`);
+                        err.status = 404;
+                        return next(err);
+                    } else {
+                        err = new Error(`Comment ${req.params.commentId} not found`);
+                        err.status = 404;
+                        return next(err);
+                    }
+                })
+                .catch(err => next(err));
+        }
     });
 
 module.exports = campsiteRouter;
